+++ CHANGELOG.md
@@ -7,3 +7,10 @@ Fixed a bug in Container Widget where if the child is not an image it wont be re
 
 ## 0.0.3
 Added README
+
+## 0.0.4
+1- Removed support for statefull widgets.
+2- Fixed a bug in container where it will redact the child without adding it to the parent container.
+
+## 0.0.5
+Added support for InkWell Widget.